requirements.txt
# Core
pyspark==3.5.4
delta-spark==3.3.0

# Jupyter
jupyterlab==4.3.6
ipykernel==6.29.5
ipywidgets==8.1.5

# Data
pandas==2.2.3
pyarrow==18.1.0
numpy==2.2.1

# Utilities
python-dotenv==1.0.1
